ERA-11397: Add support for event schemas v2 in the ReportFormSummary shown in collections
