+++ public/locales/es/details-view.json
@@ -58,6 +58,9 @@
   },
   "reportFormSummary": {
     "reportTypeLabel": "Tipo de evento",
-    "reportedByLabel": "Reportado por"
+    "reportedByLabel": "Reportado por",
+    "v2SchemaFormSummary": {
+      "collectionHumanizedValue": "{{collectionLength}} elementos"
+    }
   }
 }
