+++ src/ReportFormSummary/V2SchemaFormSummary/styles.module.scss
@@ -1,0 +1,62 @@
+@use '../../common/styles/vars/colors';
+
+.section {
+  .separator {
+    margin: 0.75rem -1rem;
+  }
+
+  .sectionLabel {
+    font-size: 1.125rem;
+    font-weight: 500;
+    margin: 0;
+  }
+
+  .columns {
+    display: flex;
+
+    .column {
+      &.fullWidth {
+        width: 100%;
+      }
+    
+      &.halfWidthLeft {
+        padding-right: 0.5rem;
+        width: 50%;
+      }
+    
+      &.halfWidthRight {
+        padding-left: 0.5rem;
+        width: 50%;
+      }
+
+      .header-large,
+      .header-medium,
+      .header-small {
+        font-weight: 500;
+        margin: 1rem 0 0.25rem;
+      }
+
+      .header-large {
+        font-size: 1.125rem;
+      }
+
+      .header-medium {
+        font-size: 1rem;
+      }
+
+      .header-small {
+        font-size: 0.875rem;
+      }
+
+      .fieldLabel {
+        color: colors.$secondary-medium-gray;
+        font-size: 0.875rem;
+        margin: 0.75rem 0 0;
+      }
+  
+      .fieldValue {
+        margin: 0;
+      }
+    }
+  }
+}
