+++ src/ReportManager/DetailsSection/SchemaForm/fields/DateTime/index.js
@@ -1,7 +1,7 @@
 import React, { memo, useEffect, useState } from 'react';
 import { format, isValid, parseISO } from 'date-fns';
 
-import { DATE_TIME_ELEMENT_INPUT_TYPES } from '../../constants';
+import { DATE_TIME_ELEMENT_INPUT_TYPES } from '../../../../../utils/v2-event-schemas/constants';
 
 import DatePicker, { isValidDate, EMPTY_DATE_VALUE } from '../../../../../DatePicker';
 import DateTimePicker, { EMPTY_DATE_TIME_VALUE } from '../../../../../DateTimePicker';
