+++ src/ReportManager/DetailsSection/SchemaForm/index.js
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
-import { FORM_ELEMENT_TYPES, ROOT_CANVAS_ID } from './constants';
-import makeFieldsFromSchema from './utils/makeFieldsFromSchema';
+import { FORM_ELEMENT_TYPES, ROOT_CANVAS_ID } from '../../../utils/v2-event-schemas/constants';
+import makeFieldsFromSchema from '../../../utils/v2-event-schemas/makeFieldsFromSchema';
 import useMapLocationMarkers from './utils/useMapLocationMarkers';
 import useSchemaValidations from './utils/useSchemaValidations';
