+++ src/utils/v2-event-schemas/constants.js
@@ -1,0 +1,40 @@
+export const CHOICE_LIST_ELEMENT_CHOICE_TYPES = {
+  EXISTING_CHOICE_LIST: 'EXISTING_CHOICE_LIST',
+  MY_DATA: 'MY_DATA',
+};
+
+export const CHOICE_LIST_ELEMENT_INPUT_TYPES = {
+  DROPDOWN: 'DROPDOWN',
+  LIST: 'LIST',
+};
+
+export const DATE_TIME_ELEMENT_INPUT_TYPES = {
+  DATE_TIME: 'DATE_TIME',
+  DATE: 'DATE',
+  TIME: 'TIME',
+};
+
+export const FORM_ELEMENT_TYPES = {
+  ATTACHMENT: 'ATTACHMENT',
+  CHOICE_LIST: 'CHOICE_LIST',
+  COLLECTION: 'COLLECTION',
+  DATE_TIME: 'DATE_TIME',
+  HEADER: 'HEADER',
+  LOCATION: 'LOCATION',
+  NUMERIC: 'NUMERIC',
+  SECTION: 'SECTION',
+  TEXT: 'TEXT',
+};
+
+export const HEADER_ELEMENT_SIZES = {
+  LARGE: 'LARGE',
+  MEDIUM: 'MEDIUM',
+  SMALL: 'SMALL',
+};
+
+export const ROOT_CANVAS_ID = 'root';
+
+export const TEXT_ELEMENT_INPUT_TYPES = {
+  SHORT: 'SHORT_TEXT',
+  LONG: 'LONG_TEXT',
+};
